Move get_user_auth to api.base.utils
