+++ api/nodes/permissions.py
@@ -1,15 +1,8 @@
-from website.models import Node, Pointer
+# -*- coding: utf-8 -*-
 from rest_framework import permissions
 
-from framework.auth import Auth
-
-def get_user_auth(request):
-    user = request.user
-    if user.is_anonymous():
-        auth = Auth(None)
-    else:
-        auth = Auth(user)
-    return auth
+from website.models import Node, Pointer
+from api.base.utils import get_user_auth
 
 class ContributorOrPublic(permissions.BasePermission):
